Last big improvements before 21X

+++ DataFormats/TrackReco/src/TrajectorySateOnDetInfo.cc
@@ -11,8 +11,8 @@ TrajectorySateOnDetInfo::TrajectorySateOnDetInfo(const LocalTrajectoryParameters
 }
 
 unsigned int TrajectorySateOnDetInfo::charge(){
- const vector<uint16_t>& Ampls       = _theCluster->amplitudes();
-//   const vector<uint8_t> &  Ampls      = _theCluster->amplitudes();
+// const vector<uint16_t>& Ampls       = _theCluster->amplitudes();
+   const vector<uint8_t> &  Ampls      = _theCluster->amplitudes();
   
    unsigned int charge=0;
    for(unsigned int a=0;a<Ampls.size();a++){charge+=Ampls[a];}
